# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sinovera Transit Global API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quote requests
      security: [{ clerkAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: number, default: 1 }
        - name: limit
          in: query
          schema: { type: number, default: 20 }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Quote requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteListResponse"

    post:
      operationId: createQuote
      tags: [quotes]
      summary: Submit a quote request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Quote request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote request
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"

    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update quote request (admin)
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"

  # ── Notifications ──────────────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security: [{ clerkAuth: [] }]
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security: [{ clerkAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security: [{ clerkAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Audit Logs ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ── Dashboard ─────────────────────────────────────────────────────────────
...
    QuoteRequest:
      type: object
      required:
        - id
        - referenceNumber
        - contactName
        - contactEmail
        - originCity
        - originCountry
        - destinationCity
        - destinationCountry
        - serviceType
        - status
        - createdAt
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        originCity:
          type: string
        originCountry:
          type: string
        destinationCity:
          type: string
        destinationCountry:
          type: string
        serviceType:
          type: string
        weightKg:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        cargoDescription:
          type: ["string", "null"]
        declaredValue:
          type: ["string", "null"]
        currency:
          type: string
        specialRequirements:
          type: ["string", "null"]
        status:
          type: string
        quotedPrice:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    QuoteListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/QuoteRequest"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
    QuoteInput:
      type: object
      required:
        - contactName
        - contactEmail
        - originCity
        - originCountry
        - destinationCity
        - destinationCountry
        - serviceType
      properties:
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        companyName:
          type: string
        originCity:
          type: string
        originCountry:
          type: string
        destinationCity:
          type: string
        destinationCountry:
          type: string
        serviceType:
          type: string
        weightKg:
          type: string
        dimensions:
          type: string
        cargoDescription:
          type: string
        declaredValue:
          type: string
        currency:
          type: string
        specialRequirements:
          type: string
    QuoteUpdate:
      type: object
      properties:
        status:
          type: string
        quotedPrice:
          type: string
        adminNotes:
          type: string

    # ── Notification ──────────────────────────────────────────────────────────
    Notification:
      type: object
      required: [id, title, message, type, channel, isRead, createdAt]
      properties:
        id:
          type: string
        customerId:
          type: ["string", "null"]
        shipmentId:
          type: ["string", "null"]
        title:
          type: string
        message:
          type: string
        type:
          type: string
        channel:
          type: string
        isRead:
          type: boolean
        readAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ── Audit Log ─────────────────────────────────────────────────────────────
...